% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bears2021}
\alias{bears2021}
\title{Locations of identified scat samples of bears}
\format{
A data frame with 715 rows and 4 variables:
\describe{
  \item{id}{Identifier of individual}
  \item{date}{Date scat sample was found}
  \item{east}{Easting in SWEREF99 TM coordinate system (meters)}
  \item{north}{Northing in SWEREF99 TM coordinate system (meters)}
}
}
\source{
\url{https://www.rovbase.se/}
}
\usage{
bears2021
}
\description{
A dataset of locations of female bear scat samples collected
during the 2011 survey in Norrbotten county, Sweden
}
\keyword{datasets}
