% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_expand_grid.R
\name{pd_expand_grid}
\alias{pd_expand_grid}
\title{Expands a grid to reduce border effect when plotting}
\usage{
pd_expand_grid(grid, coords = c("east", "north"))
}
\arguments{
\item{grid}{A grid}

\item{coords}{Names of (planar) coordinates (eastings and northings)}
}
\value{
An expanded grid
}
\description{
Mainly for internal use
}
