% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_fit.R
\name{pd_fit}
\alias{pd_fit}
\title{Fit spatial smooths (GAMs) of effort and intensity}
\usage{
pd_fit(
  data,
  region,
  id_var = "id",
  coords = c("east", "north"),
  basis_dimension = 50,
  fixed_dim = FALSE,
  cell_size = c(10000, 10000),
  min_unit = NULL
)
}
\arguments{
\item{data}{Data containing the locations of individual captures, currently the package is expecting coordinates in SWEREF 99}

\item{region}{Survey region as an sf object}

\item{id_var}{Name of individual identifier field in data}

\item{coords}{Names of (planar) coordinates (eastings and northings)}

\item{basis_dimension}{Maximum basis dimension of spatial smooths}

\item{fixed_dim}{If TRUE, a non-penalised regression spline is used}

\item{cell_size}{Width and height of grid used to aggregate coordinates for intensity smooth}

\item{min_unit}{Name of minimal unit in data. For each combination of id_var and min_unit, a maximum of one capture is recorded.}
}
\value{
An object of class "popdensity"
}
\description{
Fit spatial smooths (GAMs) of effort and intensity
}
\examples{
norrbotten <- pd_SWE_counties(county_name = "Norrbotten")
fit_2021 <- pd_fit(data = bears2021, region = norrbotten)
plot(fit_2021) + ggplot2::ggtitle("Animal density (females per square 10km)")
}
