% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_gridcounts.R
\name{pd_gridcounts}
\alias{pd_gridcounts}
\title{Counts individuals on a grid}
\usage{
pd_gridcounts(
  data,
  region,
  coords = c("east", "north"),
  cell_size = c(10000, 10000)
)
}
\arguments{
\item{data}{Data containing the locations individual activity centers, currently the package is expecting coordinates in SWEREF 99}

\item{region}{Survey region as an sf object}

\item{coords}{Names of (planar) coordinates (eastings and northings)}

\item{cell_size}{Width and height of grid used to aggregate coordinates for intensity smooth}
}
\value{
A data-frame with locations (grid centers) and counts
}
\description{
Counts individuals on a grid
}
