% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_import_rovbase.R
\name{pd_import_rovbase}
\alias{pd_import_rovbase}
\title{Cleans and renames file exported from RovBase}
\usage{
pd_import_rovbase(file)
}
\arguments{
\item{file}{Filename of RovBase Excel export}
}
\value{
A data-frame with fields date, id, species, sex, sample_type, east, north, county, collector
}
\description{
Samples without id (filled Individ) are not included
}
