% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_rootogram.R
\name{pd_rootogram}
\alias{pd_rootogram}
\title{Plots a rootogram of either the intensity or effort fit}
\usage{
pd_rootogram(fit, type = c("effort", "apparent_intensity"))
}
\arguments{
\item{fit}{Object of class popdensity}

\item{type}{Variable to be plotted}
}
\value{
A ggplot object
}
\description{
Plots a rootogram of either the intensity or effort fit
}
\examples{
norrbotten <- pd_SWE_counties(county_name = "Norrbotten")
fit_2021 <- pd_fit(data = bears2021, region = norrbotten)
# Poisson distribution of individuals per grid cell seems ok
pd_rootogram(fit_2021, type = "apparent_intensity")
# Zero-truncated Poisson distribution captures per individuals
# does not capture all individual heterogeneity
pd_rootogram(fit_2021, type = "effort")
}
