% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_SWE_counties.R
\name{pd_SWE_counties}
\alias{pd_SWE_counties}
\title{Loads sf boundaries of swedish counties in SWEREF99 TM}
\usage{
pd_SWE_counties(county_name = NA, county_nr = NA, ...)
}
\arguments{
\item{county_name}{Vector of possibly partial county names to be loaded}

\item{county_nr}{Vector of county number to be loaded}

\item{...}{Passed to st_read}
}
\value{
An sf polygon
}
\description{
Loads sf boundaries of swedish counties in SWEREF99 TM
}
