% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.popdensity.R
\name{plot.popdensity}
\alias{plot.popdensity}
\title{Generic to plot a popdensity object}
\usage{
\method{plot}{popdensity}(x, type = c("pop_density", "mean_captures", "p_capture"))
}
\arguments{
\item{x}{Object of class popdensity}

\item{type}{Variable to be plotted}
}
\value{
A ggplot object
}
\description{
Generic to plot a popdensity object
}
