% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.popdensity.R
\name{predict.popdensity}
\alias{predict.popdensity}
\title{Generic to predict a popdensity object on grid}
\usage{
\method{predict}{popdensity}(
  object,
  type = c("all", "pop_density", "mean_captures", "p_capture"),
  expand_grid = FALSE,
  ...
)
}
\arguments{
\item{object}{Object of class popdensity}

\item{type}{Variables to be predicted}

\item{expand_grid}{Should grid be expanded?}

\item{...}{Passed to mgcv::predict.gam}
}
\value{
A table with locations
}
\description{
Generic to predict a popdensity object on grid
}
