% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.popdensity.R
\name{print.popdensity}
\alias{print.popdensity}
\title{Generic to print a popdensity object}
\usage{
\method{print}{popdensity}(x, ...)
}
\arguments{
\item{x}{Object of class popdensity}

\item{...}{Not passed}
}
\value{
Basic printout
}
\description{
Generic to print a popdensity object
}
