% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.popdensity.R
\name{summary.popdensity}
\alias{summary.popdensity}
\title{Generic to summarise a popdensity object}
\usage{
\method{summary}{popdensity}(x, ...)
}
\arguments{
\item{x}{Object of class popdensity}

\item{...}{Not passed}
}
\value{
Summaries of the fitted objects
}
\description{
Generic to summarise a popdensity object
}
