% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SWEcounties}
\alias{SWEcounties}
\title{Boundaries of Swedish counties}
\format{
An sf object with polygons describing boundaries of Sweden's 21 counties
\describe{
  \item{LnKod}{County integer code}
  \item{LnNamn}{County name}
  \item{geometry}{Polygon describing county boundaries}
}
}
\source{
\url{https://www.scb.se/hitta-statistik/regional-statistik-och-kartor/regionala-indelningar/digitala-granser/}
}
\usage{
SWEcounties
}
\description{
Boundaries of Sweden's 21 counties in (planar) coordinate system SWEREF99 TM
}
\keyword{datasets}
